refactor index.js and app.js
