+++ klausur-service/app.js
@@ -1,6 +1,8 @@
 const express = require('express');
 const app = express();
 
+const session = require('express-session')
+const MongoStore = require('connect-mongo')
 
 const router = require('./routes');
 const errorHandler = require('./middleware/errorHandlingMiddleware');
@@ -8,6 +10,28 @@ const errorHandler = require('./middleware/errorHandlingMiddleware');
 //region Middleware
 app.use(express.json());
 app.use(express.urlencoded({extended: false}));
+
+const store = MongoStore.create({
+    mongoUrl: process.env.DB_URI,
+    collectionName: 'sessions'
+})
+
+const sess = session({
+    name: 'onlineklausur.sid',
+    secret: 'safesecret',
+    saveUninitialized: true,
+    resave: true, // TODO: SET TO FALSE AFTER KLAUSUR SELECTION IS IMPLEMENTED
+    store: store,
+    cookie: {
+        httpOnly: true,
+        maxAge: 999999999999,
+        sameSite: 'lax',
+        secure: false
+    }
+})
+
+app.use(sess)
+
 app.use(express.static(__dirname + "/static"));
 app.use('/api', router);
 //endregion Middleware
